pyproject.toml
[project]
name = "asistente-gastos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.0",
    "google-api-python-client>=2.185.0",
    "google-auth>=2.41.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
